requirements.txt
fastapi 
keras
torch
numpy 
pandas 
pillow 
scikit-learn
uvicorn 
